requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
anndata==0.12.10
annotated-doc==0.0.4
annotated-types==0.7.0
annoy==1.17.3
anthropic==0.83.0
anyio==4.12.1
array-api-compat==1.13.0
asciitree==0.3.3
asyncer==0.0.15
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.6
celltypist==1.7.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cuid==0.4
cycler==0.12.1
Cython==3.2.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.129.0
fasteners==0.20
filelock==3.24.3
filetype==1.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
GEOparse==2.0.4
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
groovy==0.1.2
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
harmonypy==0.2.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
igraph==1.0.0
ImageIO==2.37.2
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.10
langchain-anthropic==1.3.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langchainhub==0.1.21
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.5
Lazify==0.4.0
lazy_loader==0.4
legacy-api-wrap==1.5
leidenalg==0.11.0
literalai==0.1.201
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numcodecs==0.15.1
numpy==2.4.2
openai==2.21.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.52.4
opentelemetry-instrumentation-alephalpha==0.52.4
opentelemetry-instrumentation-anthropic==0.52.4
opentelemetry-instrumentation-bedrock==0.52.4
opentelemetry-instrumentation-chromadb==0.52.4
opentelemetry-instrumentation-cohere==0.52.4
opentelemetry-instrumentation-crewai==0.52.4
opentelemetry-instrumentation-google-generativeai==0.52.4
opentelemetry-instrumentation-groq==0.52.4
opentelemetry-instrumentation-haystack==0.52.4
opentelemetry-instrumentation-lancedb==0.52.4
opentelemetry-instrumentation-langchain==0.52.4
opentelemetry-instrumentation-llamaindex==0.52.4
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.52.4
opentelemetry-instrumentation-mcp==0.52.4
opentelemetry-instrumentation-milvus==0.52.4
opentelemetry-instrumentation-mistralai==0.52.4
opentelemetry-instrumentation-ollama==0.52.4
opentelemetry-instrumentation-openai==0.52.4
opentelemetry-instrumentation-openai-agents==0.52.4
opentelemetry-instrumentation-pinecone==0.52.4
opentelemetry-instrumentation-qdrant==0.52.4
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.52.4
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.52.4
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.52.4
opentelemetry-instrumentation-transformers==0.52.4
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.52.4
opentelemetry-instrumentation-voyageai==0.52.4
opentelemetry-instrumentation-watsonx==0.52.4
opentelemetry-instrumentation-weaviate==0.52.4
opentelemetry-instrumentation-writer==0.52.4
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
patsy==1.0.2
pillow==12.1.1
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.27.1
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
scanpy==1.11.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
scrublet==0.2.3
seaborn==0.13.2
semantic-version==2.10.0
session-info2==0.4
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
statsmodels==0.14.6
streamlit==1.54.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.4
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2026.2.16
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traceloop-sdk==0.52.4
typer==0.24.0
typer-slim==0.24.0
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zarr==2.18.7
zipp==3.23.0
zstandard==0.25.0
